[SAP] Added netapp credentials to vmdk/fcd options
This patch adds vmware_netapp_user and vmware_netapp_password
to the options for the vmdk/fcd driver.

This patch also adds a lookup for the netapp_fqdn for a
datastore.  The netapp_fqdn is a customa attribute on every
datastore for our deployment.  This makes it easy to lookup
the fqdn to connect to the right netapp that owns the datastore

Change-Id: I151b5d6d95d7801c265e47196cf15f55f27da168