[SAP] Add cinder-manage sap clean_old_lock_files command
We have a lot of orphan lock files around. At some point, there will be
too many and creating new files will fail. Therefore, we have to clean
them up.

This command lists all the files, all volumes and all snapshots. It
might take quite some time to run it.

Change-Id: Icc74c3744bedc8e2755e5c7ebe080d350401927d

SAP: Add --batch-size to clean_old_lock_files command

We can run into situations where there are millions of (orphan) lock
files around. In this case, building the list of lock files first will
take up a lot of memory. Since it doesn't matter for the correctness if
we handle all or just a subset of the lock files, as long as we retrieve
the volumes/snapshots afterwards, we now support cleaning batches of
lock files. The number of lock files to look at can be set with the
`--batch-size` parameter. It defaults to 10000.

Change-Id: I6467b32c4fe05a526a75a6f43a270f53d76e1f97