[SAP] Allow SAPFCDFilter as valid scheduler filter
In order for cinder config to be able to use the SAPFCDFilter,
it must first be defined in the setup.cfg.  This simply makes
the filter available as a choice in the config file.

Change-Id: Ib65efa32bbae5535c44456da99416734b0919096
(cherry picked from commit f273fddebd6fe4b41d5bffa90a63b5ef4d61b03d)