[SAP] Fix SAPFCDFilter
This patche squashes the following:

[SAP] Allow SAPFCDFilter as valid scheduler filter

In order for cinder config to be able to use the SAPFCDFilter,
it must first be defined in the setup.cfg.  This simply makes
the filter available as a choice in the config file.

[SAP] Fix SAPFCDFilter

This fixes an issue with the filter where the destination pool
was not passed in to the migration request.

This filter ensures a migration is allowed if:
1) the pool is different on the same vcenter.
2) if the vcenter is different the migration is only allowed on the same
   pool that the volume currently lives on. This prevents data movement.

[SAP] update SAPFCDFilter add tests

This patch updates the logic in the SAPFCDFilter to be more
permissive.  Also added unit tests to test the scenarios
of the filtering.

Change-Id: Ie2a8fc502e335963246aeec64cb5b96ba9c169d3

[SAP] Allow SAPFCDFilter for other ops (#242)

This patch allows the SAPFCDFilter to run under 2 operations
now.  The operations are 'migrate_volume' and
'find_backend_for_connector'

We want to ensure the FCD filter is run when an attach operation
kicks off a ConnectorRejected due to the volume needing to be
migrated to another vcenter.

Change-Id: I7837cd045afba42b4f05eb746bf789840b7816c0

[SAP] SAPFCDFilter checks for backend_state.host

When the admin doesn't pass in a destination host for the
migrate_volume operation, the check should go against the
backend_state.host entry.  This is the case for the
find_backend_for_connector operation as well, since there is
no destination_host specified in the request spec.

Change-Id: Iccec77bf72c4c0db3bfd48d9d7a83025c6fb92a1