[SAP] Fix SAPFCDFilter
This fixes an issue with the filter where the destination pool
was not passed in to the migration request.

This filter ensures a migration is allowed if:
1) the pool is different on the same vcenter.
2) if the vcenter is different the migration is only allowed on the same
   pool that the volume currently lives on. This prevents data movement.

[SAP] update SAPFCDFilter add tests

This patch updates the logic in the SAPFCDFilter to be more
permissive.  Also added unit tests to test the scenarios
of the filtering.

Change-Id: Ie2a8fc502e335963246aeec64cb5b96ba9c169d3
(cherry picked from commit ebee0f7f71b1cf59b1ccab374b83f7a75620b8dc)