[SAP] Allow FCD migration to down pools
This patch updates the SAPDownFilter to allow pools to pass
the filter if the pool is down if and only if the operation
on the volume is migrate and find_backend_for_connector.

This will allow a volume to be migrated from 1 vcenter to another
to the same pool/datastore when the target pool is down/drain.
It will require to removal of the pool_state extra spec in the
volume type, otherwise the capabilities filter will drop the pool.

Change-Id: I9bd6119bb3d7203bf2358c317e2699fac6730337