[SAP] Improve parallel creation from snapshot/volume
The source volume/snapshot was locked while the new volume was
creating, preventing other volumes from being created in parallel
from the same source. This was to prevent the source from being
deleted while the volume is being created from it.

We move that lock from the manager upper to the API layer until the
volume reaches the 'creating' state, then the lock is released.

The delete_snapshot / delete_volume will check the DB if there are
volumes creating from it before proceeding with the deletion.

The caveat is that, if a volume is stuck in 'creating', it will
still prevent the source from being deleted.

Change-Id: I8be26242474a2b4eba676978b533f8cad40e292b