Pin flake8-import-order<0.19.0
flake8-import-order has a new release[1] on June 20th that breaks
Cinder in a different way.

./cinder/api/api_utils.py:28:1: I300 TYPE_CHECKING block should have one newline above.
./cinder/api/common.py:32:1: I300 TYPE_CHECKING block should have one newline above.
./cinder/cmd/backup.py:48:1: I300 TYPE_CHECKING block should have one newline above.
./cinder/cmd/volume.py:52:1: I300 TYPE_CHECKING block should have one newline above.

It would be good to fix the changes but for now, it's best to pin
the flake8-import-order to <0.19.X where the jobs were stable.
We can pin it to a higher version when needed but we want to
unblock the gate as a priority for now.

[1] https://pypi.org/project/flake8-import-order/0.19.1/

Change-Id: Ic99814a61c93a9249ae9fbe5ecd5c510cb6e31ed