[SAP] Added new SAPShardRetypeMigrationFilter
The purpose of this filter is to only allow retypes
and migrations of vmdk -> fcd volumes to land on the same
shard.  We shouldn't allow vmdk -> fcd to happen accross vcenters.

[SAP] Fix declaration of ShardRetypeMigrationFilter

This last patch had the wrong file location for the new
SAPShardRetypeMigrationFilter in setup.cfg

Allow non-migration operation in SAPShardRetypeMigrationFilter
Change-Id: Iaf3852de7c6c10f7afb15e4fe897490706faeef9

[SAP] Allow migration on the same host.

This patch fixes/updates the SAPShardRetypeMigrationFilter
to allow migrating a volume to another pool on the same host.

The filter should only filter out in the case where
we are migrating from vmdk -> fcd and the shards are not the
same.

Change-Id: I40bed3ad7eb84a7d72143ce1523d47a27f53f781