Add file_exists check to file_clone (#271)
Add file_exists check to file_clone
Safeguard clone operation, that failed due to overwrite flag was set, but no file exists.
The api fails if the file does not exists and the overwrite flag is set.
The fix is checking the dest directory if the file exists, it sets the flag, but if not, that omits setting it.
Change-Id: I87c21c6b3962b590af1e636f54988d990994dec0