[SAP] NetApp client_cmode_rest fixes
Add file_exists check to file_clone (#271)

Add file_exists check to file_clone
Safeguard clone operation, that failed due to overwrite flag was set, but no file exists.
The api fails if the file does not exists and the overwrite flag is set.
The fix is checking the dest directory if the file exists, it sets the flag, but if not, that omits setting it.

Fix file_assign_qos ESCAPE special chars in file_path (#268)

fix_rename_file (#267)

You don't need to escape the "." and "/" in the json body.
The result wold be a bad filename outside the desired directory

Add function to query and set comment on NetApp volumes (#277)

Change-Id: Ic2ba4ac9e2aedd4fddecf3aaadb77fa2b7038864