[SAP] Allow migration on the same host.
This patch fixes/updates the SAPShardRetypeMigrationFilter
to allow migrating a volume to another pool on the same host.

The filter should only filter out in the case where
we are migrating from vmdk -> fcd and the shards are not the
same.

Change-Id: I40bed3ad7eb84a7d72143ce1523d47a27f53f781
(cherry picked from commit f4fe6a313452390d030fc36712593bcf745472a2)