[SAP] Add new admin API to set pool state
This patch adds a new admin API to allow an admin
to set the state of a cinder pool.  This allows
admins to set a pool to drain or available.  When a pool
is set to drain, then the pool is marked as down and the
SAPPoolDownFilter will not allow new provisioning against
that pool.

This patch also moves the custom SAP recount_host_stats API
call from the cinder builtin services contrib API to the new
SAP contrib API.  This localizes changes for SAP's code.

Change-Id: I4a2c6bb6ce178a2e3b175b2b1152a7522430af32