Cherry-pick:[vmware] Retry fetch_stream_optimized_image on DuplicateName
fetch_stream_optimized_image currently deletes the backing on any error.
The same function is used when reimaging a volume in the VolumeManager,
so when a volume with the id already exists, which causes the DuplicateName
exception to be raised and in turn not only stops the image from being
re-imaged, but the backing to be deleted.

Retrying the method in that case will result in the desired outcome
of the volume being re-created from scratch.

Signed-off-by: fabian.wiesel@sap.com
Related-Bug: 2097771
Change-Id: I7723ebcd02a7dd4b107aeab1a4c3c12389664fea