[SAP] Add version to connection_info
This patch adds a version key/value pair in the return of
initialize_connection.  This is only a metadata field for
the SAP hammer volume checker to know what is the expected
schema and values for the cinder volume attachment connection
info.  This can be used to help automatically fix known issues.

This version is added to the following SAP used drivers:
vmware/vmdk
vmware/fcd
netapp/nfs_cmode

NOTE:
Any changes to the resulting connection_info by the above drivers
need a version bump.  So future code reviews need to check that
the version is bumped if there are any changes only to the
connection_info returned by initialize_connection.

Change-Id: Ica1ea0b563d6dcc7ee30657b372ad0f369316cc0